+++ examples/Direct/sendReactionFromUserSearch.php
@@ -14,6 +14,7 @@ $truncatedDebug = false;
 
 //////////////////////
 $queryUser = 'selenagomez'; // :)
+$reaction = '🎉';
 //////////////////////
 
 $ig = new \InstagramAPI\Instagram($debug, $truncatedDebug);
@@ -184,37 +185,37 @@ try {
             ],
     ], ['module' => 'profile']);
 
-    $ig->event->sendNavigation('following', 'unified_follow_lists', 'unified_follow_lists', null, null,
+    $ig->event->sendNavigation('followers', 'unified_follow_lists', 'unified_follow_lists', null, null,
         [
-            'source_tab'    => 'following',
-            'dest_tab'      => 'following',
+            'source_tab'    => 'followers',
+            'dest_tab'      => 'followers',
         ]
     );
 
     $ig->discover->surfaceWithSu($userId);
 
     $rankToken = \InstagramAPI\Signatures::generateUUID();
-    $followings = $ig->people->getFollowing($userId, $rankToken);
+    $followers = $ig->people->getFollowers($userId, $rankToken);
 
-    $followingList = [];
-    foreach ($followings->getUsers() as $following) {
-        $followingList[] = $following->getPk();
+    $followerList = [];
+    foreach ($followers->getUsers() as $follower) {
+        $followerList[] = $follower->getPk();
     }
 
-    $ig->people->getFriendships($followingList);
+    $ig->people->getFriendships($followerList);
     $ig->discover->markSuSeen();
     $ig->discover->getAyml();
-    $ig->people->getFriendships($followingList);
+    $ig->people->getFriendships($followerList);
 
     $ig->event->sendNavigation('button', 'unified_follow_lists', 'profile');
 
-    $ig->event->sendProfileView($followingList[0]);
+    $ig->event->sendProfileView($followerList[0]);
 
-    $ig->people->getFriendship($followingList[0]);
-    $ig->highlight->getUserFeed($followingList[0]);
-    $ig->people->getInfoById($followingList[0]);
-    $ig->story->getUserStoryFeed($followingList[0]);
-    $userFeed = $ig->timeline->getUserFeed($followingList[0]);
+    $ig->people->getFriendship($followerList[0]);
+    $ig->highlight->getUserFeed($followerList[0]);
+    $ig->people->getInfoById($followerList[0]);
+    $storyFeed = $ig->story->getUserStoryFeed($followerList[0]);
+    $userFeed = $ig->timeline->getUserFeed($followerList[0]);
     $items = $userFeed->getItems();
 
     $c = 0;
@@ -226,24 +227,44 @@ try {
         $c++;
     }
 
-    $ig->event->sendNavigation('button', 'profile', 'feed_contextual_profile');
-
-    $ig->event->sendOrganicMediaImpression($items[0], 'feed_contextual_profile');
-
-    $ig->media->getCommentInfos($items[0]->getId());
-
-    $ig->event->sendOrganicNumberOfLikes($items[0], 'feed_contextual_profile');
-    $previewComments = $items[0]->getPreviewComments();
-    if ($previewComments !== null) {
-        foreach ($previewComments as $comment) {
-            $ig->event->sendCommentImpression($items[0], $comment->getUserId(), $comment->getPk(), $comment->getCommentLikeCount());
-        }
+    if ($storyFeed->getReel() === null) {
+        echo 'User has no active stories';
+        exit();
     }
 
-    // Since we are going to like the first item of the media, the position in
-    // the feed is 0. If you want to like the second item, it would position 1, and so on.
-    $ig->media->like($items[0]->getId(), 0);
-    $ig->event->sendOrganicLike($items[0], 'feed_contextual_profile', null, null, $ig->session_id);
+    $storyItems = $storyFeed->getReel()->getItems();
+    $following = $storyFeed->getReel()->getUser()->getFriendshipStatus()->getFollowing();
+    $ig->event->sendNavigation('button', 'profile', 'reel_profile');
+
+    $viewerSession = \InstagramAPI\Signatures::generateUUID();
+    $traySession = \InstagramAPI\Signatures::generateUUID();
+    $rankToken = \InstagramAPI\Signatures::generateUUID();
+
+    $ig->event->sendOrganicReelImpression($storyItems[0], $viewerSession, $traySession, $rankToken, $following, 'reel_profile');
+    $ig->event->sendOrganicMediaImpression($storyItems[0], 'reel_profile', ['story_ranking_token' => $rankToken, 'tray_session_id' => $traySession, 'viewer_session_id' => $viewerSession]);
+    $ig->story->markMediaSeen([$storyItems[0]]);
+    $ig->event->sendOrganicViewedImpression($storyItems[0], 'reel_profile', $viewerSession, $traySession, $rankToken);
+
+    $recipients = [
+        'users' => [
+            $followerList[0],
+        ],
+    ];
+    if ($storyItems[0]->getMediaType() === 1) {
+        $mediaType = 'photo';
+    } else {
+        $mediaType = 'video';
+    }
+
+    $clientContext = \InstagramAPI\Utils::generateClientContext();
+
+    $ig->direct->sendStoryReaction($recipients, $reaction, $storyItems[0]->getId(), ['client_context' => $clientContext]);
+
+    $ig->event->sendDirectMessageIntentOrAttempt('send_intent', $clientContext, 'reel_share');
+    $ig->event->sendDirectMessageIntentOrAttempt('send_attempt', $clientContext, 'reel_share');
+    $ig->event->sendDirectMessageIntentOrAttempt('sent', $clientContext, 'reel_share');
+
+    $ig->event->sendNavigation('back', 'reel_profile', 'profile');
 
     // forceSendBatch() should be only used if you are "closing" the app so all the events that
     // are queued will be sent. Batch event will automatically be sent when it reaches 50 events.