+++ examples/Like/likeFromTimeline.php
@@ -31,6 +31,17 @@ try {
     // In this case we are going to like the first item, if you are going to like any other
     // item, you should also send organic media impressios to all "viewed" medias.
     $ig->event->sendOrganicMediaImpression($item, 'feed_timeline');
+
+    $ig->media->getCommentInfos($item->getId());
+
+    $ig->event->sendOrganicNumberOfLikes($item, 'feed_timeline');
+    $previewComments = $item->getPreviewComments();
+    if ($previewComments !== null) {
+        foreach ($previewComments as $comment) {
+            $ig->event->sendCommentImpression($item, $comment->getUserId(), $comment->getPk(), $comment->getCommentLikeCount());
+        }
+    }
+
     // Since we are going to like the first item of the media, the position in
     // the feed is 0. If you want to like the second item, it would position 1, and so on.
     $ig->media->like($item->getId(), 0);